Regenerate lockfile, move it so renovate can take over
Our pipelines were failing because our lockfile was out of date and the
tzdata package changed. I'm also moving the location of the lockfile and
updating the pipeline referecnes because supposedly if it's in the root
of your repo, renovate can do this based off our rpms.in.yaml and our
repos instead of the wacky contraption I had been using to do this.
We'll see if it works :)

Signed-off-by: John Kyros <jkyros@redhat.com>