Remove broken clamav scan matrix
Clamav does not handle matrix well, see:
https://issues.redhat.com/browse/KONFLUX-9576

If we leave it in, we fail conforma because the image hashes it records
are not the ones that conforma expects and we don't figure it out until
we get to the EC stage. The only downside of this is our clam scans
can't run in parallell, but they only take seconds. Sure glad they
optimized this, eh? :P

Signed-off-by: John Kyros <jkyros@redhat.com>