Add build bumpfile for force-triggering builds
This shouldn't be necessary, but Konflux is unrelable and having to put
up a dummy PR to touch everything and force a rebuild is getting old.

This just adds a file that when poked will trigger all of the container
builds, and adds it to the on-cel of all non-catalog pipelines.

Signed-off-by: John Kyros <jkyros@redhat.com>