Fix new bundle imagerepo rewrite
So apparently now when you recreate a component, the imagerepo it
creates has a different naming scheme, the path doesn't contain the
application, so our bundle (since it was recreated) has a different path
pattern than the rest of our images.

This just updates the script to account for both/either.

Signed-off-by: John Kyros <jkyros@redhat.com>