Use protobuf 21.9 artifacts
Our protobuf compiler in the repo is too old to match what upstream is
using, and we're probably not going to get an update this release (we're
stuck on 3.14, we need 3.15). So we either have to use a newer compiler,
or we have to mangle args on our old compiler in our actual KEDA code to
make this work.

There are official release artifacts upstream for the protobuf compiler
and it looks like they are in use elsewhere in openshift for cases like
this, so this moves us to use the artifacts for this release.

Later when we get an update we can move back to the RPMs.

Signed-off-by: John Kyros <jkyros@redhat.com>