Move back to RPMs for artifacts
There was a bug that made generic artifacts work without exceptions, and
I think it will take us a little while to get exceptions, so I added
some repos to our activation key and found all the openshift clients, so
I think we should be able to flip back now.

I did have to wrap the protobuf compiler in the dockerfile because 3.14
(the newest compiler we can get here) needs additional options to build
v3 optional.

Signed-off-by: John Kyros <jkyros@redhat.com>